[Bug #164119535] Finished fixing tests

+++ app/api/v2/views/offices/offices_view.py
@@ -1,13 +1,25 @@
 from flask import Blueprint, jsonify, request
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.api.v2.models.offices.office_model import OfficeModel
 
 office_view = Blueprint("office_view", __name__)
 
 
+def checkIfAuthorized():
+    user = get_jwt_identity()
+    if user["role"] == "False":
+        return False
+    return True
+
+
 @office_view.route("", methods=["POST"])
 @jwt_required
 def createOffice():
+    if checkIfAuthorized() is False:
+        return jsonify({
+            "error": "You are not authorized to create an office",
+            "status": 403
+        }), 403
     response = OfficeModel(request.get_json(force=True)).createOffice()
     return jsonify(response), response["status"]
 
@@ -29,6 +41,11 @@ def getSpecificOffice(id):
 @office_view.route("/<id>", methods=["PATCH"])
 @jwt_required
 def editSpecificOffice(id):
+    if checkIfAuthorized() is False:
+        return jsonify({
+            "error": "You are not authorized to edit an office",
+            "status": 403
+        }), 403
     response = OfficeModel(request.get_json(force=True), id).editSpecificOffice()
     return jsonify(response), response["status"]
 
@@ -36,6 +53,11 @@ def editSpecificOffice(id):
 @office_view.route("/<id>", methods=["DELETE"])
 @jwt_required
 def deleteSpecificOffice(id):
+    if checkIfAuthorized() is False:
+        return jsonify({
+            "error": "You are not authorized to delete an office",
+            "status": 403
+        }), 403
     response = OfficeModel(None, id).deleteSpecificOffice()
     return jsonify(response), response["status"]
 
@@ -43,6 +65,11 @@ def deleteSpecificOffice(id):
 @office_view.route("/<id>/register", methods=["POST"])
 @jwt_required
 def userRegisterToOffice(id):
+    if checkIfAuthorized() is False:
+        return jsonify({
+            "error": "You are not authorized to register a user to office",
+            "status": 403
+        }), 403
     response = OfficeModel(request.get_json(force=True), id).userRegisterToOffice()
     return jsonify(response), response["status"]
